requirements.txt
Django==5.2.1
djangorestframework==3.16.0
matplotlib==3.10.3
sympy==1.14.0